+++ tanjun/clients.py
@@ -1907,11 +1907,13 @@ class Client(tanjun.Client):
         return self
 
     def with_listener(
-        self, event_type: type[hikari.Event], /
+        self, *event_types: type[hikari.Event]
     ) -> collections.Callable[[_ListenerCallbackSigT], _ListenerCallbackSigT]:
         # <<inherited docstring from tanjun.abc.Client>>.
         def decorator(callback: _ListenerCallbackSigT, /) -> _ListenerCallbackSigT:
-            self.add_listener(event_type, callback)
+            for event_type in event_types or utilities.infer_listener_types(callback):
+                self.add_listener(event_type, callback)
+
             return callback
 
         return decorator